browser: history: fix queries for when DQS disabled
Recently the FreeBSD port of sqlite has DQS feature disabled by default.
Without this feature enabled, it's not allowed to use double quotes
for string literals. As such quoting is used in the history.py module,
qutebrowser fails to work on such configurations.

The fix is to use single quotes instead.

ref: #7596