tests: add some text to another test HTML file
This one is for 6.5. Interesting that it was fine on 6.4.

I haven't gone and added text to all the test file because I'm hoping
it'll be something that upstream addresses in short order. So I'm only
adding text to files used by failing tests.

See d413b87c3f6 #7621 and #7624