qt6 mypy: Disable warn-unused-ignores
The ignores needed between Qt 5 and Qt 6 differ.
We could buy into e.g. only Qt 6 linting, but apparently e.g. VS Code
also shows more errors when removing the Qt 5 type ignores.

Instead, disable this for now. We might want to re-enable it when we see
a major change in the mypy changelog and filter the results manually.

81 -> 50 errors