Make sure ipc.shutdown() can be called twice
We still want to make sure that IPC is shut down early when using :restart.
However, doing it again should be a nop when quitting qutebrowser then
shuts it down again.

This accidentally worked before 6373959 ("Always delay stage 2 shutdown, not
only with prompts"), because most operations in shutdown() just happened to be
idempotent. Because we never returned to the Qt mainloop between the first and
the second call, the self._server.deleteLater() call didn't actually delete
anything yet.

After that change, however, we do end up in shutdown() again with an actually
deleted C++ QLocalServer object, so trying to call .close() on it again failed
with a RuntimeError.

Fixed #7681