walk_module workaround for PyInstaller
Restored from dd2fc8e10bb9d4a1bd0158110173793a18736d6b

Now that we are putting our data files in the qutebrowser/ directory
again pkgutil/importlib is getting confused by that dir existing and
returning us a FileLoader for `qutebrowser.components`, I think that's
what's happening anyway. Should try reverting that and this commit and
see if extensions get loaded right again.

So bring back this workaround of using the toc on the PyInstaller loader
to get the list of component modules for now.

ref: #7803