child event filter: Ignore non-QQuickWidget children
When pressing buttons on some websites, or when starting to drag, it looks like
the WebView gets new QObject children which are not actually their focus proxy.

So far, this wasn't a big issue: We only ended up installing the tab event
filter on objects where it doesn't belong.

However, with the new focus workaround from #7820, we then ended up calling
`.setFocus()` on those QObjects, causing an AttributeError.

Thus, just don't do anything if we get new children that are not actually a
QQuickWidget.

Fixes #7831