Rename Chromium setting for colors.webpage.darkmode.threshold.text
As Chromium 99.0.4785.0 (translates to Qt 6.4) renamed `TextBrightnessThreshold` to
`ForegroundBrightnessThreshold`, we need to reflect that in our mapping of the
qutebrowser option `colors.webpage.darkmode.threshold.text` to the Chromium setting.

By adding a new method (`_Definition.copy_replace_setting()`) we can introduce a new
`_Definition` for Qt 6.4 based on the preceding one for Qt 6.3, which contains the
renamed setting.

References:
* https://chromium-review.googlesource.com/c/chromium/src/+/3344100
* https://chromium-review.googlesource.com/c/chromium/src/+/3226389