Disable accelerated 2d canvas by default.
Ideally we would only enable this if we detected Intel graphics. But
that kind of feature detection is difficult to do cross platform. It's
easy to do with Qt (version.py) ... but it needs a QApplication and we
are trying to figure out what CLI args to pass to Qt, so we can't do
that.
So just disable it for everyone for now.

TODO:
* tests
* change to tri-state option, auto by default and make auto look at the
  webengine version and on for Qt6 < 6.7.0