reduce reliance of a couple of tests on an actual timer
These tests got quite flaky after my changes to implement a per message
timeout. For now I've changed them to clear messages explicitly instead
of waiting for a timer to expire.

These tests should not be relying on a real timer, why make your tests
slow like that? They should be mocking out the timer and firing the
signal where appropriate.

My changes to how messages are cleared has previously made some of the
other tests more complicated. In those tests I mocked `time.time()` and
used that to manage when `update()` cleared messages. But since `time()`
is called is called in multiple places (for "now" and
"message._created_at") I don't think that the most intuitive mechanism.

Alternatives:
1. provide created_at to messages when creating them (or just make it an
   incrementing int) to have more control over it
2. and/or have an `expire_message()` helper that'll make sure a message
   gets cleared on next `update()`

Either way it might be nicer for the tests to add private
`_get_message_created_at()` and or `_get_now()` helper methods on the
messageview that can be mocked to be able to control the two usages of
`time.time()` separately.