Get extension loading working without load_module()
Importlib says this about load_module:

> Deprecated since version 3.4: The recommended API for loading a module
> is exec_module() (and create_module()). Loaders should implement it
> instead of load_module(). The import machinery takes care of all the
> other responsibilities of load_module() when exec_module() is
> implemented.

I'm not 100% sure how to implement the new way of doing things with this
setup, I've got something working now but I'm not sure it is the best
way of doing it.

To recap, what I'm trying to do here is load extensions from the
extension folder (eg `$XDG_DATA_DIR/qutebrowser/extensions/`) under a
namespace we control (eg `qutebrowser.extensions.third_party`). I don't
remember why I set it up like that but there is probably a good reason.

What I was doing previously was creating a module and not loading it (I
think?) in walk_extensions, by extensions putting it in sys.modules,
then later load_component would call the normal import mechanism that
would find it already loaded.

Now I can't find a way to create a module without importing it. Eg
`spec.loader.create_module(spec)` just returns None, which the docs say
means "use the standard import mechanism". Not a very helpful
replacement for `load_module()`!

But the loader returned by the finder returned by walk_packages does now
how to import the module. So for now I'm saving the whole spec for
`walk_component()` so it can re-use that loader. And then after creating
the module I have to add it to sys.modules manually and then exec it.
Having to manipulate sys.modules manually is problematic and the biggest
sign that I'm doing the wrong thing here.

I suspect that the proper fix is to register a finder in `sys.meta_path`
that'll handler that fake "third_part" package so that you can just to
`import qute.third_party.my_extension` and it'll map the module name to
a path and create a module from a file somehow. The finder I see here is
`FileFinder('/real/path/to/extensions/')` so that might be a good
starting spot. Links for how to create Finders:

https://docs.python.org/3/library/importlib.html#setting-up-an-importer
https://stackoverflow.com/questions/57126085/how-to-use-sys-meta-path-with-python-3-x