Fix showing widget when newly enabled
In the initial implementation the widget had `show()` and `hide()`
called on it and we used `isVisible()` to tell whether we should update
the widget on events.

Since the widget showing in the statubar got refactored in https://github.com/qutebrowser/qutebrowser/pull/6448/files
the only other widgets that have an `on_tab_changed()` method don't get
get `.show()` called on them anymore, than handle that themselves with
the help of a `.enabled` attribute. Not 100% sure if it's the right
thing to do to lump this settings widget in with them just because it
also takes a tab (eg do the history and progress widgets handle showing
themselves because they hide when they aren't needed? The setting widget
is always visible when configured) but it's not big deal.

In general I want to refactor the statusbar to not have explicit
knowledge of the widgets it can show (eg have a common abstract class
for them with same defaults and helpers). In the meantime this is all
pretty self contained.

More words on that eventual refactoring here: https://github.com/qutebrowser/qutebrowser/issues/7407