Make "new_chrome" UA quirk a bit clever
We only want to be faking a newer chrome version, if we are already on a
newer one, we don't want to be faking an older one!

Another scenario is that we might want to fake a perpetually newer
version for some site. But that seems like a very problematic scenario
and I hope that never comes up.