lint: allow comparing variables to 0
With pylint 3 there is a new option: https://pylint.readthedocs.io/en/latest/user_guide/messages/convention/use-implicit-booleaness-not-comparison-to-zero.html

It's disabled by default but we enable all warnings and disabled them as
desired.

This one is of the opinion that:

    if x == 0:

is bad and

    if x:

is good.

I feel that the first one (x == 0) is more clear. We aren't checking for
truthiness here, we are checking for a literal value, its very
intentional.

One might argue that being precious about making the type here is
redundant in current year with type checking tooling and all that. But
there are like a hundred of these checks in the code base so it seems a
well established pattern anyhow.

In summary, the new warning doesn't have a very strong use case and we
would prefer to stick with out established pattern.