Fix resource pak file location on mac.
On CI I'm seeing:

    No such file or directory: '/Users/runner/work/qutebrowser/qutebrowser/.tox/py39-pyqt515/lib/python3.9/site-packages/PyQt5/Qt5/resources'

Downloading the PyQt6_WebEngine_Qt6 wheel for mac from PyPI I can see the pak
file is at:

    ./PyQt6/Qt6/lib/QtWebEngineCore.framework/Resources/qtwebengine_resources.pak

And on a qutebrowser install made by pyinstaller it is at (symlinks in curly
braces):

    /Applications/qutebrowser.app/Contents/{Resources,Frameworks}/PyQt6/Qt6/lib/QtWebEngineCore.framework/{Resources,Versions/A/Resources}/qtwebengine_resources.pak

And the Qt data path for reference is:

    /Applications/qutebrowser.app/Contents/Frameworks/PyQt6/Qt6

So it looks like essentially we need to add a "lib/QtWebEngineCore.framework"
in there and capitalise "Resources".

A bit annoying to have the special case and hardocde paths like this. But it
should be pretty stable? I had a look at importlib_resources to try to see if
it could fine this stuff for us but I don't think it can.

I've checked with the pyinstaller windows builds and with the windows wheel
from PyPI and neither seem to need any special handling.