help mypy deal with pathlib
Seems mypy and I agree on something. It thinks that pathlib is abusing
the division operation and making python a less intuitive
language. Not sure why it is complaining though really, doing `Path /
"one" / "two"` seems to work fine in practice, and its on the pathlib
documentation page.

I was seeing this error locally and on CI

   qutebrowser/misc/pakjoy.py:155: error: Unsupported left operand type for / ("str")  [operator]
               resources_dir /= "lib" / "QtWebEngineCore.framework" / "Resources"