Override getters for some WebEngineView attributes
With PyQt6-WebEngine 6.6.0 some pointer return types are now wrapped in
Optionals[]. In practice they should never be none though. So in a low effort
way of keeping the types we have to deal with in the calling code to
what they were before I'm overriding these getters that just do `assert thing
not None` to tell mypy not to worry about that case.

    QWebEngineView.page()
      The docs don't say anything but the source says it should always return
      something: https://github.com/qt/qtwebengine/blob/6.6.0/src/webenginewidgets/api/qwebengineview.cpp#L1001
    QWebEngineView.history()
      calls QWebEnginePage.history()
      where it is always set in the constructor: https://github.com/qt/qtwebengine/blob/6.6.0/src/core/api/qwebenginepage.cpp#L90
    QWebEngineView.settings()
      calls QWebEnginePage.settings()
      where it is always set in the constructor: https://github.com/qt/qtwebengine/blob/6.6.0/src/core/api/qwebenginepage.cpp#L90

ref: https://github.com/qutebrowser/qutebrowser/pull/7990